Completely new parser
This is a total rewrite of the original code, splitting the parsing up into three layers - there's a THJSONInputStream, which homogenizes the input stream across different OS line endings and provides a peek-ahead facility and counts line numbers and columns. There's then a THJSONTokenizer, that returns valid categorized THJSON tokens in the order they are parsed - again with a peek ahead facility.  Finally there's the THJSONReader, that reads a stream of THJSON tokens from the tokenizer and turns that stream into a series of events.

In the /lib folder there's a Notepad++ syntax grammar thing that colours THJSON files reasonably well. It could be a lot better but Notepad++ syntax grammars are a bit limited.